requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
appnope==0.1.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.3.0
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
comm==0.2.2
debugpy==1.8.11
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.1.0
Flask==3.1.0
Flask-WTF==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kombu==5.4.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
multidict==6.1.0
nest-asyncio==1.6.0
numpy==1.23.0
openai==0.28
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.18.3
